got rid of warning
Former-commit-id: 895141c72ca6c0eb9843403538f85f78138ff6a6

+++ libIRDB/include/core/type.hpp
@@ -112,7 +112,7 @@ class AggregateType_t : public Type_t
 		void AddAggregatedType(Type_t *t, int pos);
 		virtual int GetNumAggregatedTypes() const { return refTypes.size(); } 
 		Type_t* GetAggregatedType(int pos) const { 
-			return (pos >= 0 && pos < refTypes.size()) ? refTypes.at(pos) : NULL;
+			return (pos >= 0 && pos < (int)refTypes.size()) ? refTypes.at(pos) : NULL;
 		}
 
 		std::string toString() {
